Split download actions modal into components
